Changed the way a histogram is read from CSV import.
Using the standard C++ way of reading in a line into several variables using the string stream makes problems and often returns 0 for the second value and not the actual value in the file.

Now using atof on substrings that were read manually. This can lead to errors (that still have to be caught) if one reads in a file with three or more columns.